backend/pyproject.toml
[project]
name = "ask-docs-backend"
version = "0.1.0"
description = "Local-LLM RAG backend for documentation Q&A"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "PyMuPDF>=1.24.0",
    "beautifulsoup4>=4.12.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
